requirements.txt
streamlit
openai
python-dotenv
PyMuPDF
tqdm
google-generativeai
plotly